added support for arbitrary maps and some map projections
I included the map information stuff into the main pipeline, so that
specifying a region or set of countries with the --region keyword now
appropriately selects a region. Alternatively, a convex hull around the sampled
points is supported, or an intersection of any options. I also added very
basic support for some map projections, but passing arguments for the
projections is not really possible under the current scheme, it might be
necessary to add those as a functional argument.